chatbot/requirements.txt
gradio
google-genai
PyPDF2
numpy
scikit-learn
chromadb
sentence-transformers